+++ packages/forms/src/directives/reactive_directives/form_group_name.ts
@@ -13,7 +13,7 @@ import {NG_ASYNC_VALIDATORS, NG_VALIDATORS} from '../../validators';
 import {AbstractFormGroupDirective} from '../abstract_form_group_directive';
 import {ControlContainer} from '../control_container';
 import {ReactiveErrors} from '../reactive_errors';
-import {composeAsyncValidators, composeValidators, controlPath} from '../shared';
+import {controlPath} from '../shared';
 import {AsyncValidator, AsyncValidatorFn, Validator, ValidatorFn} from '../validators';
 
 import {FormGroupDirective} from './form_group_directive';
@@ -91,8 +91,8 @@ export class FormGroupName extends AbstractFormGroupDirective implements OnInit,
           (AsyncValidator|AsyncValidatorFn)[]) {
     super();
     this._parent = parent;
-    this._validators = validators;
-    this._asyncValidators = asyncValidators;
+    this._setValidators(validators);
+    this._setAsyncValidators(asyncValidators);
   }
 
   /** @internal */
@@ -137,12 +137,6 @@ export class FormArrayName extends ControlContainer implements OnInit, OnDestroy
   /** @internal */
   _parent: ControlContainer;
 
-  /** @internal */
-  _validators: (Validator|ValidatorFn)[];
-
-  /** @internal */
-  _asyncValidators: (AsyncValidator|AsyncValidatorFn)[];
-
   /**
    * @description
    * Tracks the name of the `FormArray` bound to the directive. The name corresponds
@@ -162,8 +156,8 @@ export class FormArrayName extends ControlContainer implements OnInit, OnDestroy
           (AsyncValidator|AsyncValidatorFn)[]) {
     super();
     this._parent = parent;
-    this._validators = validators;
-    this._asyncValidators = asyncValidators;
+    this._setValidators(validators);
+    this._setAsyncValidators(asyncValidators);
   }
 
   /**
@@ -209,23 +203,6 @@ export class FormArrayName extends ControlContainer implements OnInit, OnDestroy
    */
   get path(): string[] {
     return controlPath(this.name == null ? this.name : this.name.toString(), this._parent);
-  }
-
-  /**
-   * @description
-   * Synchronous validator function composed of all the synchronous validators registered with this
-   * directive.
-   */
-  get validator(): ValidatorFn|null {
-    return composeValidators(this._validators);
-  }
-
-  /**
-   * @description
-   * Async validator function composed of all the async validators registered with this directive.
-   */
-  get asyncValidator(): AsyncValidatorFn|null {
-    return composeAsyncValidators(this._asyncValidators);
   }
 
   private _checkParentType(): void {
